+++ CloudImage-backend/CloudImage/Service/IApiKeyService.cs
@@ -5,5 +5,5 @@ public interface IApiKeyService
     bool IsValidApiKey(string apiKey);
     void AddApiKey(string apiKey);
     double GetRemainingStorage(string apiKey);
-    void UpdateUsedStorage(string apiKey, long uploadedFileSize);
+    void UpdateUsedStorage(string apiKey, double uploadedFileSize);
 }
